pyproject.toml
[project]
name = "peft-playground"
version = "0.1.0"
description = "PEFT playground with LoRA and FLoRA adapters"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch",
    "transformers>=4.39.0",
    "datasets>=2.16.0",
    "evaluate>=0.4.0",
    "accelerate>=0.25.0",
    "pyyaml>=6.0",
    "numpy>=1.23",
    "wandb>=0.15",
    "scipy",
    "sklearn"
]

[project.scripts]
peft-playground = "peft_playground.cli:main"
